Initial implementation of new playback

+++ src/key_handler/mod.rs
@@ -22,8 +22,8 @@ const X: KeyModifiers = KeyModifiers::NONE;
 const S: KeyModifiers = KeyModifiers::SHIFT;
 const C: KeyModifiers = KeyModifiers::CONTROL;
 
-const SEEK_SMALL: usize = 5;
-const SEEK_LARGE: usize = 30;
+const SEEK_SMALL: u64 = 5;
+const SEEK_LARGE: u64 = 30;
 const SCROLL_MID: usize = 5;
 const SCROLL_XTRA: usize = 20;
 const SIDEBAR_INCREMENT: isize = 1;
@@ -36,8 +36,8 @@ pub enum Action {
     TogglePause,
     PlayNext,
     PlayPrev,
-    SeekForward(usize),
-    SeekBack(usize),
+    SeekForward(u64),
+    SeekBack(u64),
 
     // Queue & Playlist Actions
     QueueSong,
